+++ EngineSource/Engine/Console.h
@@ -3,16 +3,28 @@
 #include <Engine/TypeEnun.h>
 #include <vector>
 
+/**
+* @file
+*/
+
+/**
+* @brief
+*/
 namespace Console
 {
-	// Initialized default ConVars and Commands
+	/// Initialized default ConVars and Commands
 	void InitializeConsole();
 
-	// Executes the given command string.
-	// This function will return true if the command string was evaluated successfully. Otherwise it will return false.
+	/**
+	* @brief
+	* Executes the given command string.
+	* 
+	* @return
+	* True if the command string was evaluated successfully.
+	*/
 	bool ExecuteConsoleCommand(std::string Command);
 
-	// ConVar struct.
+	/// Contains information about a console variable.
 	struct Variable
 	{
 		std::string Name;
@@ -24,7 +36,7 @@ namespace Console
 		Variable() {}
 	};
 
-	// Command struct.
+	/// Contains information about a console command.
 	struct Command
 	{
 		std::string Name;
@@ -42,25 +54,25 @@ namespace Console
 		Command() {}
 	};
 
-	// Registers a ConVar that can be modified through the console.
+	/// Registers a ConVar that can be modified through the console.
 	void RegisterConVar(Variable Var);
-	// Registers a Command that can be called through the console.
+	/// Registers a Command that can be called through the console.
 	void RegisterCommand(Command NewCommand);
 
-	// Gets the arguments passed to the most recent command. This is totally necessary, trust me.
+	/// Gets the arguments passed to the most recent command.
 	std::vector<std::string> CommandArgs();
 
 	enum ConsoleLogType
 	{
-		E_INFO = 0,
-		E_WARNING = 1,
-		E_ERROR = 2
+		Info = 0,
+		Warn = 1,
+		Err = 2
 	};
 
-	// Logging function for the console. It is basically just Log::Print() but with preset colors.
-	void ConsoleLog(std::string Message, ConsoleLogType Severity = E_INFO);
+	/// Logging function for the console. It is like Log::Print() but with preset colors.
+	void ConsoleLog(std::string Message, ConsoleLogType Severity = Info);
 
-	// A function to print out a vector of command arguments. Useful for displaying info about a command
-	// or why a certain command has invalid arguments passed to it.
-	void PrintArguments(std::vector<Command::Argument> args, ConsoleLogType Severity = E_INFO);
+	/// A function to print out a vector of command arguments. Useful for displaying info about a command
+	/// or why a certain command has invalid arguments passed to it.
+	void PrintArguments(std::vector<Command::Argument> args, ConsoleLogType Severity = Info);
 }