+++ EngineSource/Math/Physics/JoltPhysics.cpp
@@ -159,7 +159,6 @@ public:
 #endif // JPH_EXTERNAL_PROFILE || JPH_PROFILE_ENABLED
 };
 
-auto a = EMotionType::Kinematic;
 class ObjectVsBroadPhaseLayerFilterImpl : public ObjectVsBroadPhaseLayerFilter
 {
 public: